kern/vfs: perform search through every fs as a last resort
As a last resort, if a mountpoint was not found, vfs_read
should peform a search through every filesystem to see
if we can find the file anywhere. We don't have complicated
or massive filesystems right now to make this have a impact
on performance. Also simplified read_readme since we can now
read and write through the vfs instead of the respective filesystem
drivers.

Signed-off-by: vlprokopenko <vlprokopenko@disroot.org>